[fix] Compilation in Ubuntu 18.04

+++ src/VNLBayes/VideoNLBayes.cpp
@@ -691,8 +691,8 @@ unsigned estimateSimilarPatches(
 			int ct0 = ct[dt - dir];
 			float cx_f = cx0 + (use_flow ? (dir > 0 ? fflow(cx0,cy0,ct0,0) : bflow(cx0,cy0,ct0,0)) : 0.f);
 			float cy_f = cy0 + (use_flow ? (dir > 0 ? fflow(cx0,cy0,ct0,1) : bflow(cx0,cy0,ct0,1)) : 0.f);
-			cx[dt] = std::max(0., std::min((double)sz.width  - 1, round(cx_f)));
-			cy[dt] = std::max(0., std::min((double)sz.height - 1, round(cy_f)));
+			cx[dt] = std::max(0.f, std::min((float)sz.width  - 1, round(cx_f)));
+			cy[dt] = std::max(0.f, std::min((float)sz.height - 1, round(cy_f)));
 			ct[dt] = qt;
 		}
 		else
